+++ crates/bevy_pbr/src/render/mesh_bindings.rs
@@ -1,5 +1,6 @@
 //! Bind group layout related definitions for the mesh pipeline.
 
+pub use bevy_material::render::MeshLayouts;
 use bevy_math::Mat4;
 use bevy_mesh::morph::MAX_MORPH_WEIGHTS;
 use bevy_render::{
@@ -158,43 +159,6 @@ mod entry {
             resource: BindingResource::SamplerArray(samplers),
         }
     }
-}
-
-/// All possible [`BindGroupLayout`]s in bevy's default mesh shader (`mesh.wgsl`).
-#[derive(Clone)]
-pub struct MeshLayouts {
-    /// The mesh model uniform (transform) and nothing else.
-    pub model_only: BindGroupLayoutDescriptor,
-
-    /// Includes the lightmap texture and uniform.
-    pub lightmapped: BindGroupLayoutDescriptor,
-
-    /// Also includes the uniform for skinning
-    pub skinned: BindGroupLayoutDescriptor,
-
-    /// Like [`MeshLayouts::skinned`], but includes slots for the previous
-    /// frame's joint matrices, so that we can compute motion vectors.
-    pub skinned_motion: BindGroupLayoutDescriptor,
-
-    /// Also includes the uniform and [`MorphAttributes`] for morph targets.
-    ///
-    /// [`MorphAttributes`]: bevy_mesh::morph::MorphAttributes
-    pub morphed: BindGroupLayoutDescriptor,
-
-    /// Like [`MeshLayouts::morphed`], but includes a slot for the previous
-    /// frame's morph weights, so that we can compute motion vectors.
-    pub morphed_motion: BindGroupLayoutDescriptor,
-
-    /// Also includes both uniforms for skinning and morph targets, also the
-    /// morph target [`MorphAttributes`] binding.
-    ///
-    /// [`MorphAttributes`]: bevy_mesh::morph::MorphAttributes
-    pub morphed_skinned: BindGroupLayoutDescriptor,
-
-    /// Like [`MeshLayouts::morphed_skinned`], but includes slots for the
-    /// previous frame's joint matrices and morph weights, so that we can
-    /// compute motion vectors.
-    pub morphed_skinned_motion: BindGroupLayoutDescriptor,
 }
 
 pub trait MeshLayoutsBuilders {
